Animate the indicator on the Other screen
